Adicao de Convenio, HistoricoClinico, Prescricao, Anamnese, Cids, valições, DTOS, correção de bugs, e erros de lógica + arquivos para adicionar todos os cids no banco de dados

+++ src/main/java/com/clinica/service/RelatorioService.java
@@ -36,7 +36,7 @@ public class RelatorioService {
 
 				Month mes = pagamento.getDataPagamento().getMonth();
 
-				faturamento.put(mes, faturamento.getOrDefault(mes, 0.0) + pagamento.getValor());}}
+				faturamento.put(mes, faturamento.getOrDefault(mes, 0.0) + pagamento.getValor().doubleValue());}}
 
 		return faturamento;
 	}
